Split ipc into main and renderer
